OLCS-14261 Fix issue if DigitalSignature was null

See merge request !279

+++ app/api/module/Api/src/Domain/CommandHandler/Application/UpdateApplicationCompletion.php
@@ -83,8 +83,10 @@ final class UpdateApplicationCompletion extends AbstractCommandHandler implement
             $application->setDeclarationConfirmation('N');
             $application->setSignatureType(null);
             $digitalSignature = $application->getDigitalSignature();
-            $application->setDigitalSignature(null);
-            $this->getRepo('DigitalSignature')->delete($digitalSignature);
+            if ($digitalSignature !== null) {
+                $application->setDigitalSignature(null);
+                $this->getRepo('DigitalSignature')->delete($digitalSignature);
+            }
         }
 
         $sectionsToUpdate = $this->getSectionsToUpdate($command, $completion);
